+++ lib/util/folly_iterable_extension.dart
@@ -1,0 +1,24 @@
+///
+///
+///
+extension FollyIterableExtension<T extends Enum> on Iterable<T> {
+  ///
+  ///
+  ///
+  T byNameOrDefault(
+    String? name, {
+    required T defaultValue,
+  }) {
+    if (name == null) {
+      return defaultValue;
+    }
+
+    for (final T value in this) {
+      if (value.name == name) {
+        return value;
+      }
+    }
+
+    return defaultValue;
+  }
+}